requirements.txt
streamlit>=1.28.0
openai>=1.6.0
python-dotenv>=1.0.0
pillow>=9.0.0
requests>=2.28.0